Added: `region-geogrid-params`
I want to remove `geogrids`s usage for internal storage b/c they're
not memory efficient and so it's limiting the applications ability to
injest large data sets

Ideally all data will be stores in the data matrix. Geogrids will be
used for extracting subregions and for generating SVGs. So they're
going to be just a convenient intermediary format

One primary issue is that geogrids that are being used as an
intermediary format initialize their params by using a prexisting
geogrid and then just overwritting the data

We can get the params explicited and save those in the "context" and
then use them explicitely when creating new geogrids