Added: `locations/regions` map
Regions are now referenced by key.
This makes life easier..

I can use the key name in writing to the output folder

Also litering a whole namespace with defs is prolly uglier...
How about if I want to find all the locations and make a list?
With a map I just call keywords or whatever