Update: make above and below groupings explicit (for coloring)
Before the algo would split the groups
One group would get assigned brown and the other teal
But the two groupings seemingly could get swapped at random.
It was deterministic, but depended on the dataset being run.
Now I've explicitely set the above to be teal,
and the below to be brown