Added: debug directory output
Debug output it streamed into a subfolder based on the region name.
This allows me to track output of different regions

Previously the debug output would clobber the previous run.
Which meant I needed to rerun regions all the time
and it was annoying to compare regions

The only downside is that region folders can be "stale"
and need to be updated manually after say a bugfix