Update: removed geogrids as a cached intermediary representation
The only cache should be the data matrix (+ metadata)
and maybe the SVD of it...
It's a compact represenation of all the data we need

We also need to define geogrids with a param list.
The previous method was copied the params of an existing geogrid,
which was kinda messy.

Not grid-params are a cached vector of values

Geogrids can now be reconstructed from the data matrix as needed.
Keeping all the geogrids around just eats memory and was inelegant