Added: functionality to do 1-EOF analysis
Looking at distributions and deviations off EOF1.
The variance seems to change with the strength of EOF1.
Currently plotting histograms for the variation of the grid values,
relative the the EOF1 "background"

Still need to plot this on an XY plot...
With EOF1 intensity vs variance.

That's the next step for tomorrow